# specs/012-video-transcription/contracts/seminar-recording-api.openapi.yaml
openapi: 3.1.0
info:
  title: Aither-Hemera Seminar Recording Integration
  version: 1.0.0
  description: >-
    Service contracts for validated booking lookup, durable transcription workflow
    state, and authorized short-lived media access. Hemera owns workflow records
    and booking authorization; Aither owns AssemblyAI, MUX, and Blob credentials.
servers:
  - url: https://aither.example.invalid
    description: Aither service (placeholder)
  - url: https://hemera.example.invalid
    description: Hemera service (placeholder)
...
paths:
...
  /api/service/seminar-document-access:
    post:
      tags: [Aither Service API]
      summary: Mint short-lived media access after Hemera booking authorization
      description: >-
        Called only by Hemera server-side after checking the authenticated user's
        ownership of bookingId. This endpoint is not a substitute for Hemera's
        user authorization check.
      security:
        - AitherServiceKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [bookingId, recordingId]
              properties:
                bookingId:
                  type: string
                recordingId:
                  type: string
      responses:
        '200':
          description: Temporary bearer URLs for a ready workflow
          content:
            application/json:
              schema:
                type: object
                required: [muxPlaybackUrl, transcriptUrl, muxExpiresAt, transcriptExpiresAt]
                properties:
                  muxPlaybackUrl:
                    type: string
                    format: uri
                    description: Fresh signed HLS URL; expiry exceeds the full video duration. This request-time URL is distinct from the stable muxPlaybackUrl stored in Hemera and must not be persisted or logged.
                  transcriptUrl:
                    type: string
                    format: uri
                    description: Private Blob GET URL expiring after exactly 300 seconds.
                  muxExpiresAt:
                    type: string
                    format: date-time
                  transcriptExpiresAt:
                    type: string
                    format: date-time
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Workflow is not ready for participant access
...
  /api/service/seminar-recordings/{recordingId}:
    delete:
      tags: [Aither Service API]
      summary: Revoke and delete all provider artifacts after Hemera booking/participation deletion
      description: >-
        Hemera revokes participant access before calling this endpoint. Aither
        deletes any AssemblyAI transcript/source media, MUX asset, final and
        source-staging private Blob objects, and any residual local recording
        artifact, then reports completion so Hemera
        can purge its non-visible tombstone.
      security:
        - AitherServiceKey: []
      parameters:
        - $ref: '#/components/parameters/RecordingId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [bookingId, deletionId]
              properties:
                bookingId:
                  type: string
                deletionId:
                  type: string
                  description: Hemera-generated idempotency key for this delete operation.
                deletionReason:
                  type: string
                  enum: [booking_deleted, participation_deleted, operator_abandoned]
                  description: Operator abandonment may be requested only by an authorized Hemera administrator.
      responses:
        '202':
          description: Access revoked; asset deletion accepted or already in progress
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '200':
          description: Every known provider artifact is deleted or idempotently confirmed as already deleted; safe to purge the Hemera tombstone.
...
components:
  securitySchemes:
    HemeraApiKey:
      type: apiKey
      in: header
      name: X-API-Key
    AitherServiceKey:
      type: apiKey
      in: header
      name: X-Aither-Service-Key
    AssemblyAIWebhookSecret:
      type: apiKey
      in: header
      name: X-AssemblyAI-Webhook-Secret
    ClerkSessionCookie:
      type: apiKey
      in: cookie
      name: __session
  parameters:
...
    RecordingId:
      name: recordingId
      in: path
      required: true
      schema:
        type: string
...
  responses:
    Unauthorized:
      description: Missing or invalid service credential
    NotFound:
      description: Booking or recording workflow not found